Move add logic to child components - mw
