+++ web/src/App.jsx
@@ -10,12 +10,6 @@ import './App.css';
 const App = () => {
     const userEmail = 'example@email.com';
     const [todos, setTodos] = useState([]);
-    const [data, setData] = useState({
-        user_email:  'example@email.com',
-        title: '',
-        progress: 0,
-        date: new Date(),
-      });
     const [isEditing, setIsEditing] = useState(false);
     const [currentTodo, setCurrentTodo] = useState({});
 
@@ -26,35 +20,6 @@ const App = () => {
             setTodos(response)
         } catch (err) {
           console.log(err)
-        }
-    }
-
-    const handleAddInputChange = (e) => {
-        const { value } = e.target
-        setData((data) => ({
-        ...data, title: value
-        }))
-        console.log(data)
-    };
-
-        
-    const handleAddFormSubmit = async (e) => {
-        e.preventDefault()
-        try {
-            if( data.title !== '') {
-                const response = await axios.post('http://localhost:8000/todos/create', {
-                    data: data,
-                },{
-                    headers: {
-                        'Content-Type': 'application/json'
-                    }
-                })
-                getTodos()
-                setData( { title: '' } )
-                console.log('add form submit response:', response)
-            }
-        } catch (error) {
-            console.log(error)
         }
     }
         
@@ -71,7 +36,7 @@ const App = () => {
                     setCurrentTodo={setCurrentTodo}
                 />
             ) : (
-                <AddForm todo={data} onAddInputChange={handleAddInputChange} onAddFormSubmit={handleAddFormSubmit} />
+                <AddForm getTodos={getTodos} />
             )} 
             <List sx={{ width: '100%', maxWidth: 480, bgcolor: 'background.paper' }}>
                 { todos.map((todo) => (
