+++ web/src/components/EditForm.jsx
@@ -1,5 +1,4 @@
 /* eslint-disable react/prop-types */
-// import { useState } from 'react';
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
